9614: Rename test for clarity, remove argument that isnt being used anymore
